==> Part1A_Processing_Posts.ipynb <==
# Project Overview
In this notebook and subsequent notebooks I will attempt to predict mechanically meaningful labels associated with some fan-generated roleplaying game (RPG) content.

I collected post and comment data from Reddit in multiple runs using a command-line tool I wrote for this purpose (see: [PRAW-CoDiaLS](https://github.com/nkuehnle/praw-codials))

The actual text I will be attempting to predict is hosted on one of two websites: [GM Binder](https://gmbinder.com/) or [Homebrewery](https://homebrewery.naturalcrit.com/), which are tools/content hosting providers for fan-made RPG content, primarily Dungeons and Dragons 5th Edition (DnD5e).

The labels make use of the user-assigned "flair" from the subreddit [/r/UnearthedArcana](https://www.reddit.com/r/UnearthedArcana/). UnearthedArcana describes itself as a source of "homebrew" (fan-generated) content and its community rules ensure that every post has a meaningful label ("flair") associated with it. These labels describe the "type" of game content that has been created, be it a character "class" (think achetypal high-fantasy characters: knights, wizards, archers, etc) or a "race" (elves, dwarves, humans, etc).

At the start of this document we have only the data about the Reddit submission including the submission flair link to the service described above.

%load_ext jupyter_black
# Utility Imports
import os
from pathlib import Path

# Imports for data processing/handling
import pandas as pd
from urllib.parse import urlparse
import itertools

# This Notebook
This notebook will clean up the individual links and their associated labels that were collected from Reddit.

Below, I describe some over-arching issues with the links collected within the dataset and start to fix them.

# Define a constant (CWD) to use
CWD = Path(os.getcwd())
DATA = CWD / "data"
PRAW = DATA / "praw_data"

for f in [DATA, PRAW]:
    fname = str(f)
    currname = str(CWD)
    fname = fname.replace(currname, "")
    try:
        f.mkdir(parents=True, exist_ok=False)
    except FileExistsError:
        pass
    else:
        print(f"Folder {fname} was created")

# Data Sources
There are two files (CommentsRaw.csv and SubmissionsRaw.csv)

If you are not familiar with Reddit, when users post it can take two forms, "Submissions" and "comments."

Submissions are associated with a specific community or "subreddit" and users can comment on Submissions to discuss them or provide additional commentary.

Submissions can come in multiple forms: link, text, image, etc. The submissions here are only link submissions. Comments are searched for all other types of submissions as users will frequently post an image version of their 
content and then link to the text copy in the comments on these subreddits. The vast majority of content on /r/UnearthedArcana comes in this second form (an image submission with the link in the comments).

In any case, PRAW-CoDiaLS is written so that they both share the same overall structure, although certain fields are only used for comments (and blank otherwise).


### Shared Fields

| column name | description |
| ----- | ----- |
| link | Raw text URL |
| submission_author | Reddit user who posted the comment |
| submission_id | A unique ID used for locating the submission on Reddit |
| submission_title | The title of the submission on Reddit |
| subreddit | The name of the Subreddit the content was found on (in this case, either UnearthedArcana or DnDHomebrew) |
| submission_flair | the figure number as indicated in the manuscript |
| submission_score | The approximate number of net positive votes received by the submission on Reddit, the actual values are obfuscated by Reddit to avoid manipulation |
| submission_upvote_ratio | The ratio of positive to negative votes |
| submission_date | Date on which the submission was originally made |

### Comment-only fields

| column name | description |
| ----- | ----- |
| comment_author | Reddit user who posted the comment |
| comment_id | A unique ID used for locating the comment on Reddit |
| comment_score | The approximate number of net positive votes received by the comment on Reddit, the actual values are obfuscated by Reddit to avoid manipulation |
| comment_body | Text of the comment containing markdown elements |
| comment_date | Date on which the comment was originally posted |

### Other Notes
There are many cases where matching a URL and flair are not entirely straightforward. I will discuss this as it comes up below.

subs_path = PRAW / "SubmissionsRaw.csv"
subs_df = pd.read_csv(subs_path, sep=",")  # Load in submission data
subs_df["submission_flair"] = subs_df["submission_flair"].replace("Feature", "Feat")
subs_df = subs_df.drop_duplicates(subset=["link", "submission_id"])
subs_df = subs_df.reset_index(drop=True)
subs_df.to_csv(subs_path, index=False)  # Save data
subs_df.head()

cmts_path = PRAW / "CommentsRaw.csv"
cmts_df = pd.read_csv(cmts_path, sep=",")  # Load comment data
cmts_df["submission_flair"] = cmts_df["submission_flair"].replace("Feature", "Feat")
cmts_df = cmts_df.drop_duplicates(subset=["link", "comment_id"])
cmts_df = cmts_df.reset_index(drop=True)
cmts_df.to_csv(cmts_path, index=False)  # Save data
cmts_df.head()

# Exploring random comments...

Let's print out some (7) random comments to see how they look...
 
Note several issues apparent here. I'll work through them in order in the following cells.

cmt_sample = cmts_df.sample(n=7, random_state=7)
content = zip(cmt_sample["submission_title"], cmt_sample["comment_body"])

for i, (title, body) in enumerate(content, start=1):
    print(f"\033[91mEntry {i}: " + title + "\033[0m")
    print(f"{body}\n")

### Fixing differences in URL formatting

At times URLs are represented diferently (http vs https).

To help make sure all URLs are formatted identically, I wrote a simple function.

`conform_url()` formats all URLs to https://<no www.>domain.extension/path

from src import conform_url

cmts_df["link"] = cmts_df["link"].apply(lambda x: conform_url(x))
subs_df["link"] = subs_df["link"].apply(lambda x: conform_url(x))

index_key = cmts_df.copy()
mapper = index_key["submission_flair"].to_dict()

# Issue #1: Varying link types

We can see we have some different varieties of GMBinder/Homebrewery links, including those with /share/ and /edit/ directories.

There are also some from unrelated domains (Google Docs, Wikipedia articles, etc.) and some direct links to PDF files.

For now we will be dropping any links that aren't directly to 

We also have two varieties of link text:


### URLs in raw text form
URLs with no special formating around them.
> <pre>Gmbinder Link: https://www.gmbinder.com/share/-MW5xKc1QPj4TXEgBzQm</pre>

### Markdown links
These take the form of a display text portion in brackets immediately followed by a URL in parenthesis.
> <pre>[Display text](url)</pre>

In some cases, the display text is also a url and in at least one case early on the url is incorrect in the display text portion, but correct in the actual url segment of the link code.

> <pre>[https://homebrewery.naturalcrit.com/share/MhTK9\_7zHuB6](https://homebrewery.naturalcrit.com/share/MhTK9_7zHuB6)</pre>


Since the url portion is more likely to be correct, we'll go exclusively with raw links and the url portion of the markdown code.

To do this we can use a set of regular expressions. After a few trial and errors, I arrived at the following:

#### #Homebrewery
> <pre>?:https?:\/\/)?(?:www\.)?homebrewery\.naturalcrit\.com\/[\w\-][\w\-]*\/?[\w\-]*\/?[\w\-]*</pre>

#### GMBinder
> <pre>(?:https?:\/\/)?(?:www\.)?gmbinder\.com\/[\w\-][\w\-]*\/?[\w\-]*\/?[\w\-]*[\s\)]</pre>

### Historical Note
The code block below and regular expression described above document a legacy process for handling links within comments. PRAW-CoDiaLS now generates better expressions to automatically collect them.

```
link_re = r'[\s\(](?:https?:\/\/)?(?:www\.)?homebrewery\.naturalcrit\.com\/[\w\-][\w\-]*\/?[\w\-]*\/?[\w\-]*[\s\\)]|[\s\(](?:https?:\/\/)?(?:www\.)?gmbinder\.com\/[\w\-][\w\-]*\/?[\w\-]*\/?[\w\-]*[\s\)]'
cmts_df['link'] = cmts_df['comment_body'].str.findall(link_re)
```

To keep PRAW-CoDiaLS more general-purpose, I still filter certain links below.

bad_substrings = [
    "pdf",  # Remove links to PDFs (GMBinder)
    "profile",  # Remove links to GMBinder user profiles
    "edit",  # Remove links to the main text editor/content creation tool (Homebrewery)
    "print",  # Remove links to PDFs (Homebrewery)
    "user",  # Remove links to Homebrewery user profiles
    "png",  # Image posts
    "-LKYVliROX20cBbSI0r3",  # GMBinder style example
    "-L0WUSjJyFHG7PDyIFta",  # GMbinder tutorial
    "-Kx5i9fuxhuci6BUrbJq",  # GMBiner Watercolor tutorial
    "-MbAAe7gUkgvdAdE47qp",  # Collection of links to Bamstacks' variants
    "B1Hju_QaTx",  # Homebrewery style guide
    "HyWImuH06x",  # Homebrewery styling guide
]
discarded_subs = None
discarded_cmts = None

for substring in bad_substrings:
    cmts_to_drop = cmts_df["link"].str.contains(substring)
    subs_to_drop = subs_df["link"].str.contains(substring)

    discard_dfs = (cmts_df[cmts_to_drop].copy(), subs_df[subs_to_drop].copy())
    discard_dfs[0]["discard_reason"] = "Invalid/Missing Content Type"
    discard_dfs[1]["discard_reason"] = "Invalid/Missing Content Type"

    if (len(discard_dfs[0]) > 0) & isinstance(discarded_cmts, pd.DataFrame):
        discarded_cmts = pd.concat([discarded_cmts, discard_dfs[0]])
    else:
        discarded_cmts = discard_dfs[0]

    if isinstance(discarded_subs, pd.DataFrame) and len(discard_dfs[1]) > 0:
        discarded_subs = pd.concat([discarded_subs, discard_dfs[1]])
    else:
        discarded_subs = discard_dfs[1]

    cmts_df = cmts_df[~cmts_to_drop]
    subs_df = subs_df[~subs_to_drop]

print(f"{len(discarded_cmts)} comments with invalid URL types discarded")
print(f"{len(discarded_subs)} submissions with invalid URL types discarded")

# Remove links w/o paths (i.e. just to homebrewery.naturalcrit.com or gmbinder.com)
def missing_path(url: str) -> bool:
    return urlparse(url).path == ""


missing_paths = cmts_df["link"].apply(lambda x: missing_path(x))

new_discards = cmts_df[missing_paths].copy()
new_discards["discard_reason"] = "Missing path"
print(f"{len(new_discards)} comments discarded for missing URL path")
discarded_cmts = pd.concat([discarded_cmts, new_discards])
cmts_df = cmts_df[~missing_paths]

## Removing known problematic comments based on authorship

Some users will share their related but not necessarily identically-labeled content. Thus, we should only be looking at comments where the submission and comment are made by the same author. Otherwise the flair may not necessarily apply. We can manually check these posts later.


# Make sure that the common author is the same as the post author to help ensure that they refer to the same content.
different_author = cmts_df["submission_author"] != cmts_df["comment_author"]
different_author = cmts_df[different_author].copy()
print(
    f"{len(different_author)} comments with a different author from the primary submission dropped"
)
cmts_df = cmts_df[~cmts_df.index.isin(different_author.index)]

author_issues_path = PRAW / "AuthorIssues.csv"

if author_issues_path.is_file():
    author_issues_df = pd.read_csv(author_issues_path, sep=",", index_col="idx")

    # Check if any links are new.
    links_to_add = different_author[
        ~different_author.index.isin(author_issues_df.index)
    ]
    print(
        f"Merged {len(links_to_add)} links into existing set of {len(author_issues_df)}"
    )

    # Add only new links
    author_issues_df = pd.concat([author_issues_df, links_to_add])
    author_issues_df["manually_reviewed"] = author_issues_df[
        "manually_reviewed"
    ].fillna(False)

else:
    author_issues_df = different_author
    author_issues_df["manually_reviewed"] = False

print(
    f"Saving {len(author_issues_df)} comment links where the author differs from the submission for manual review"
)
print(
    f"{author_issues_df['manually_reviewed'].sum()} out of {len(author_issues_df)} comments with authorship issues have been reviewed already."
)
author_issues_df.to_csv(author_issues_path, index_label="idx")

# Issue #2: Lack of 1:1 correspondence between links and submissions
For several reasons, a given piece of user content can be associated with multiple submissions/comments and vice versa.
I'll go through them one by one below.

#### General Notes:
* The main idea here is to have a nice variety of documents with meaningful flair, so I'll always preserve UnearthedArcana links when present and never remove an UnearthedArcana link in favor of a DnDHomebrew one.
* I'm going to be keeping track of the rows I drop in case I decide to manually inspect for QC later on.

##### All the functions defined below for filtering out repeat links will <u>**modify** the main dataframe(s) **inplace**</u> and <u>**return** any discarded rows</u>.
Individual functions described and imported below:
* `filter_simple_cmt_issues()`
* `get_multi_link_parents_by_author()`
* `filter_cmt_multi_links_by_date()`

### Issue #2: Part A -- Simple Issues (obvious redundancies)

There are two obvious types of easily resolvable redundancies I can identify.

1. Submissions where a link is posted in the comments multiple times.
2. Links collected from comments, where a meaningfully flaired primary submission was also found

This functionality is implemented in the `filter_simple_cmt_issues()` function


from src import filter_simple_cmt_issues

new_discards = filter_simple_cmt_issues(cmts_df, subs_df)
discarded_cmts = pd.concat([discarded_cmts, new_discards])

### Issue #2: Part B -- Cross-posts
Some content has been shared on both /r/DnDHomebrew and /r/UnearthedArcana.

Since /r/UnearthedArcana has useful flair and /r/DnDHomebrew does not, we will always prioritize /r/UnearthedArcana links.

A helper function for this, `filter_dh_cross_posts()`, has been defined.

# Import function to remove links posted to /r/DnDHomebrew if they are already available from /r/UnearthedArcana
from src import filter_dh_cross_posts

new_discards = filter_dh_cross_posts(cmts_df, subs_df)
discarded_cmts = pd.concat([discarded_cmts, new_discards[0]])
discarded_subs = pd.concat([discarded_subs, new_discards[1]])

### Issue #2: Part C -- Multi-link comments
Frequently an author will self-promote their other content within a comment. We'll need to identify and resolve these situations somehow later on.

For instance take the comment with ID *fkzbprj* under the submission titled "Sword Mage - 4e's teleporting arcane defender brought back as an Artificer subclass - v2.1" where the author, *cyberhawk92*, links to a grand total of **14** other piees of content they've created.

I've defined two functions to help with the process of cleaning up some of these cases.

`get_multi_link_parents_by_author()` creates a dictionary with authors as keys and lists of all posts by that author as values. Only authors with >1 post are included.

`filter_cmt_multi_links_by_date()` goes through and discards comments where the link has been repeated but was already posted in an older comment (more likely to be the original)

from src import filter_cmt_multi_links_by_date

newer_multi_cmts = filter_cmt_multi_links_by_date(cmts_df)
discarded_cmts = pd.concat([discarded_cmts, newer_multi_cmts])
cmts_df = cmts_df[~cmts_df.index.isin(newer_multi_cmts.index)]
newer_multi_cmts.sort_values(by="link")

### Issue #2: Part D -- Exporting the remaining multi-link comments
I'm going to export a csv of the comments w/ multiple links for a manual review.

from src import get_multi_link_parents_by_author

multi_link_dict = get_multi_link_parents_by_author(cmts_df)
offenders = [v for v in multi_link_dict.values()]
print(f"{len(offenders)} remaining authors with repeat posts")

offenders = list(itertools.chain(*offenders))
multi_link_df = cmts_df[cmts_df["submission_id"].isin(offenders)]
cmts_df = cmts_df[~cmts_df["submission_id"].isin(offenders)]
multi_link_df

multi_link_path = PRAW / "MultiLinkCmts.csv"

if multi_link_path.is_file():
    _multi_link_df = pd.read_csv(multi_link_path, index_col="idx")
    print(_multi_link_df.index.isin(cmts_df.index).sum() / len(_multi_link_df))
    missing_cmts = multi_link_df[~multi_link_df.index.isin(_multi_link_df.index)]
    print(
        f"Merged {len(missing_cmts)} links into existing set of {len(_multi_link_df)}"
    )
    multi_link_df = pd.concat([_multi_link_df, missing_cmts])
    multi_link_df["manually_reviewed"] = multi_link_df["manually_reviewed"].fillna(
        False
    )
else:
    multi_link_df["manually_reviewed"] = False
multi_link_df = multi_link_df.sort_index()

print(
    f"Saving {len(multi_link_df)} links from comments with mutliple links for manual review."
)
print(
    f"{multi_link_df['manually_reviewed'].sum()} out of {len(multi_link_df)} links from comments with multiple associated links have been reviewed already."
)
multi_link_df.to_csv(multi_link_path, index_label="idx")

### Historical Note
Some links were previously manually reviewed for being part of "multi link comments" but changes in the pipeline and/or collection of additional led to the other links associated with these comments being filtered out earlier or otherwise changing their status.

Nonetheless, I'd rather go with the manually reviewed labels since comments can still be mis-associated with submission flair and this adds an additional layer of certainty for even more labels.

print(
    f"Dropped {cmts_df.index.isin(multi_link_df.index).sum()} additional comments already processed in mutli-link comment file"
)
cmts_df = cmts_df[~cmts_df.index.isin(multi_link_df.index)]

Un-discard any manually reviewed content (i.e. in previous iterations of dataset present in the multi-link file).

Check why they were discarded.

discard_reasons = discarded_cmts[discarded_cmts.index.isin(multi_link_df.index)][
    "discard_reason"
].value_counts()
discarded_cmts = discarded_cmts[~discarded_cmts.index.isin(multi_link_df.index)]
discard_reasons

discarded_cmts.to_csv(PRAW / "CommentsDropped.csv", index_label="idx")
cmts_df.to_csv(PRAW / "CommentsFiltered.csv", index_label="idx")

### Filtering redundant primary submissions
Instances where a link has been submitted more than once. In these cases use the most recent /r/UnearthedArcana version since it's most likely to be correct.

subs_to_drop = subs_df.sort_values(
    by=["link", "subreddit", "submission_date"], axis=0, ascending=[True, False, False]
)
subs_to_drop = subs_to_drop[
    subs_to_drop.duplicated(subset=["link", "submission_flair"], keep="first")
]
subs_to_drop["discard_reason"] = "Repeated link with newer submission"
discarded_subs = pd.concat([discarded_subs, subs_to_drop])

subs_df = subs_df[~subs_df.index.isin(subs_to_drop.index)]
remaining_dups = subs_df.duplicated(subset="link", keep=False).sum()
print(f"Dropped {len(subs_to_drop)} duplicated links submitted as primary submissions")
print(f"{remaining_dups} duplicate submissions remaining")

Let's inspect the remaining duplicate submission links

subs_df[subs_df.duplicated(subset="link", keep=False)]

subs_df[subs_df.duplicated(subset="link", keep=False)]["link"].values[0]

Both of these appear to be somewhat accurate. It's a compendium of classes (which has grown since the initial post). I'll keep the compendium link.

subs_df = subs_df[subs_df.index != 823]

discarded_subs.to_csv(PRAW / "SubmissionsDropped.csv", index_label="idx")
subs_df.to_csv(PRAW / "SubmissionsFiltered.csv", index_label="idx")
