==> Part1D_Finalizing_Data.ipynb <==
# Project Overview
In this notebook and subsequent notebooks I will attempt to predict mechanically meaningful labels associated with some fan-generated roleplaying game (RPG) content.

I collected post and comment data from Reddit in multiple runs using a command-line tool I wrote for this purpose (see: [PRAW_FILES-CoDiaLS](https://github.com/nkuehnle/praw-codials))

The actual text I will be attempting to predict is hosted on one of two websites: [GM Binder](https://gmbinder.com/) or [Homebrewery](https://homebrewery.naturalcrit.com/), which are tools/content hosting providers for fan-made RPG content, primarily Dungeons and Dragons 5th Edition (DnD5e).

The labels make use of the user-assigned "flair" from the subreddit [/r/UnearthedArcana](https://www.reddit.com/r/UnearthedArcana/). UnearthedArcana describes itself as a source of "homebrew" (fan-generated) content and its community rules ensure that every post has a meaningful label ("flair") associated with it. These labels describe the "type" of game content that has been created, be it a character "class" (think achetypal high-fantasy characters knights, wizards, archers, etc) or a "race" (elves, dwarves, humans, etc).

%load_ext jupyter_black
# Utility Imports
import os
from pathlib import Path

# Imports for data processing/handling
import numpy as np
import pandas as pd
import pickle as pkl

# NLP-specific processing
from nltk.tokenize import word_tokenize
from langdetect import detect

# Custom modules
from src.preprocessing import (
    fix_dt,
    data_io,
    get_section_df,
    sections_df_to_docs,
    EmbeddingAwareTokenizer,
)
from src.eda_utils import get_char_counts, count_stopwords, calc_token_sizes

# Bringing it all together

At this point we've:
1. Collected links containing fan-generated Dungeons and Dragons game content posted by Reddit users to [/r/UnearthedArcana](https://reddit.com/r/unearthedarcana) or [/r/DnDHomebrew](https://reddit.com/r/dndhomebrew) using my [PRAW_FILES-CoDiaLS](https://github.com/nkuehnle/praw-codials) tool
2. Filered out or semi-manually reviewed links to get the largest set of unique, accurately "flaired" (labeled) links
3. Scraped this content from [one](https://homebrewery.naturalcrit.com) of [two](https://gmbinder.com) websites that host this content (as literal string markdown text with some CSS elements)
4. Cleaned as much irrelevant information from the texts (i.e. styling elements and so forth)

### In this notebook I'll:
1. Ensure column datatypes are properly set
2. Filter once more for a few potential problems that might have been missed
3. Load the text as either sections or full documents into two possible dataframes
4. Get word and sentence tokens at both the section and document level (this takes a while)

# Define a constant (CWD) to use
CWD = Path(os.getcwd())
DATA = CWD / "data"
PRAW_FILES = DATA / "praw_data"
TEXT_FILES = DATA / "text_files"
CLEAN_TEXT_FILES = DATA / "clean_text_files"
CORPUS_FILES = DATA / "corpus_files"
CORPUS_FILES.mkdir(parents=True, exist_ok=True)

metadata_path = CORPUS_FILES / "Metadata.pkl"
metadata: pd.DataFrame = pd.read_pickle(metadata_path)
metadata = metadata.sort_values(by="UID")
metadata.info()

# Update manually reviewed /r/DnDHomeBrew content
At the previous notebook I generated a list of labels from /r/DnDHomebrew (which has optional and/or less meaningful flair) to review.

Here I'll incorporate any new information I've gathered from manually reviewing these texts

label_review_path = PRAW_FILES / "LabelsToReview.csv"
# Read in reviewed labels
reviewed_labels = pd.read_csv(label_review_path)
reviewed_labels["manually_reviewed"] = reviewed_labels["manually_reviewed"].fillna(
    False
)
manually_reviewed = reviewed_labels["manually_reviewed"]

reviewed_labels = reviewed_labels[manually_reviewed].copy()
reviewed_labels["submission_flair"] = reviewed_labels["corrected_flair"]

# Update values
to_update = metadata["UID"].isin(reviewed_labels["UID"])
mapper = reviewed_labels.set_index("UID")["corrected_flair"].to_dict()
metadata.loc[to_update, "submission_flair"] = metadata.loc[to_update, "UID"].map(mapper)

print(f"{to_update.sum()} /r/DnDHomebrew entries updated.")

### Set column datatypes

# Dates
metadata["submission_date"] = metadata["submission_date"].apply(lambda x: fix_dt(x))
metadata["comment_date"] = metadata["comment_date"].apply(lambda x: fix_dt(x))
# Boolean
metadata["manually_reviewed"] = metadata["manually_reviewed"].fillna(False)
metadata["manually_reviewed"] = metadata["manually_reviewed"].astype("boolean")
# All primary submission links are related to the submission...
metadata.loc[metadata["comment_id"].isna(), "related_link"] = True
metadata["related_link"] = metadata["related_link"].astype("boolean")
# Numbers
metadata["comment_score"] = metadata["comment_score"].astype("Int32")
metadata["submission_score"] = metadata["submission_score"].astype("Int32")
metadata["submission_upvote_ratio"] = metadata["submission_upvote_ratio"].astype(
    "float32"
)
metadata["UID"] = metadata["UID"].astype("Int16")
# Categorical
metadata["subreddit"] = metadata["subreddit"].astype("category")
metadata["submission_flair"] = metadata["submission_flair"].astype("category")

### Remove any remaining duplicates (now that DnDHomebrew content has flair)

dups = metadata.sort_values(
    by=["manually_reviewed", "subreddit", "submission_date", "related_link"],
    axis=0,
    ascending=[False, False, True, False],
)
dups = dups[dups.duplicated(subset=["src_url", "submission_flair"], keep="first")]
metadata = metadata[~metadata["UID"].isin(dups["UID"])]

Confirming no src_urls have been duplicated with differing flair

metadata[metadata["src_url"].duplicated(keep=False)].sort_values("src_url")

### Remove any invalid/missing content
A few bad links have slipped through at some point. These will be dropped below.

#### Manual Review/missing flair
After manual review, some links still had no relevant flair, seen and dropped below.

metadata[metadata["submission_flair"].isna()]["link"].values
metadata = metadata[~metadata["submission_flair"].isna()]

#### Missing/invalid content
Likely collected in a previous pass or made evident after cleaning the markdown
In a few cases, non-English texts are present. Due to the small number, we won't attempt training/translating.

metadata["raw_markdown"] = metadata["UID"].apply(
    lambda x: data_io.read_src_txt_from_file(x, TEXT_FILES)
)
metadata["clean_markdown"] = metadata["UID"].apply(
    lambda x: data_io.read_src_txt_from_file(x, CLEAN_TEXT_FILES)
)

to_drop = [
    # Drop GMBinder links that are dead.
    metadata["raw_markdown"].str.contains("NoSuchKey"),
    # Drop Homebrwery links that are dead
    metadata["raw_markdown"].str.contains("This content has been moved"),
    # Some are not found
    metadata["raw_markdown"].str.contains("Error: File not found"),
    # There are a few examples of nonsensical text, which is generally short, thus I'm doing a check for anything less than 150 characters
    metadata["raw_markdown"].str.len() <= 200,
    metadata["clean_markdown"].str.len() <= 200,
    # Drop anything totally empty.
    metadata["raw_markdown"] == "",
    metadata["clean_markdown"] == "",
]

drop_mask = None
for mask in to_drop:
    if isinstance(drop_mask, pd.Series) or isinstance(drop_mask, np.ndarray):
        drop_mask = drop_mask | mask
    else:
        drop_mask = mask
print(f"Dropped {sum(drop_mask)} texts based on absence of content:")
print(*metadata[drop_mask]["link"], sep="\n")
dropped = metadata[drop_mask].copy()
metadata = metadata[~drop_mask]

# Check language
metadata["lang"] = metadata["clean_markdown"].map(detect)
not_english = metadata["lang"] != "en"
print(f"{not_english.sum()} non-english texts removed:")
print(*metadata[not_english]["link"], sep="\n")
dropped = pd.concat([dropped, metadata[not_english]])
metadata = metadata[~not_english]

print(f"{len(metadata)} texts remain")

dropped[dropped["lang"] == "en"]

dropped[dropped["lang"] != "en"]

### Remove rare flair

metadata.groupby(by=["submission_flair"])["UID"].agg("count")

metadata.groupby(by=["submission_flair"])["UID"].agg("count")
MAJOR_LABELS = [
    "Class",
    "Item",
    "Monster",
    "Race",
    "Spell",
    "Subclass",
    "Feat",
    "Compendium",
    "Mechanic",
    "Background",
]
metadata[~metadata["submission_flair"].isin(MAJOR_LABELS)]

metadata = metadata[metadata["submission_flair"].isin(MAJOR_LABELS)].copy()
metadata["submission_flair"] = metadata[
    "submission_flair"
].cat.remove_unused_categories()

metadata = metadata.sort_values("UID")
metadata.info()

# Create Corpus Dataframes & Tokenize
### Calculation of QC/Language Metrics
Metrics are calculated at the sentence/word token and full string levels on:
1. the raw markdown
2. the cleaned markdown
3. the section-level text
Metrics for markdown are calculated first and use simple tokenization to get some basic information for EDA/QC purposes.
Final/real-word implementation will not bother with this except as part of continual learning/QC

**Speed:** This takes ~1-2 minutes for each phase.

### Get section-level text
This involves using the NLTK markdown corpus reader to split the text up into sections. The plan is for section-level data to mostly be used for cleaning out things like credit sections and/or unsupervised learning to get better segmented labels for supervised learning cases (i.e. multi-class sequence to sequence models)

**Speed:** akes approx. ~4 minutes (~1.5 min to initialize NLTK markdown corpus reader, ~2.5 min to get section-level data--most of that is labeling the sections as credit/non-credit)

### Custom word tokenization

My tokenizer here will attempt to resolve a number of ambiguities that exist in the form of spelling variations in a smart way that maximizes the number of words taken from a pre-existing set of word embeddings. In this case, we'll use the GLoVe (Wikipedia+Gigaword) embeddings taken from Gensim. **For now we only need the list of vocabulary terms.** In the event that a word is unknown and rare within the text corpus, it will use the part-of-speech tag from NLTK as an approximation of the meaning.

My custom tokenizer is slower than standard word tokenization largely due to its effort to retain as many tokens in a meaningful way as possible. The **entire** text that is being tagged is passed into the PoS tagger if an unknown term is encountered since performance is generally best within the full context of the surrounding tokens.

**Note**: Section tokenization is *particularly* slow. It's ~10x faster to tokenize by document, but we lose information that defines sections if we tokenize first. A better future implementation would be to figure out where the start and end of each section by looking for the indices of key tokens surrounding each token/text. Since I only need to do this once and going from section tokens -> document tokens is trivial, I implemented it like this for now.

**Speed:** Fitting the tokenizer takes ~1-2 minutes and mostly cross references uncommon or hyphen/other delimited characters with "known words" that are either highly common in the corpus or are present in the word embeddings' vocabulary.
This tokenization is a bit overboard for bag-of-words models, but I would like to implement some sequence-to-sequence stuff later on.

sec_df_path = DATA / "section_corpus.pkl"
tokenizer_path = DATA / "tokenizer.pkl"

if sec_df_path.is_file():
    # No longer need metadata, free up memory and import section df
    del metadata
    section_df = pd.read_pickle(sec_df_path)
    with open(tokenizer_path, "rb") as p:
        tokenizer = pkl.load(p)
else:
    print("Calculating markdown-level QC metrics")
    # Calculate basic markdown QC/lang metrics
    get_char_counts(metadata, pref="raw_md", text_col="raw_markdown")
    get_char_counts(metadata, pref="clean_md", text_col="clean_markdown")
    count_stopwords(
        metadata,
        pref="clean_md",
        text_col="clean_markdown",
        tokenizer=word_tokenize,
        text_to_lower=True,
    )
    calc_token_sizes(metadata, pref="clean_md", text_col="clean_markdown")

    # Get sections
    print("Getting sections-level dataframe.")
    section_df = get_section_df(metadata, CLEAN_TEXT_FILES)  # See function/module for details
    del metadata

    # Get tokens
    if tokenizer_path.is_file():
        print("Loading tokenizer")
        with open(tokenizer_path, "rb") as p:
            tokenizer = pkl.load(p)
    else:
        print("Fitting tokenizer")
        import gensim.downloader as api

        # Load WV for embedding vocab
        vectors = api.load("glove-wiki-gigaword-100")
        vector_vocab = [k for k in vectors.key_to_index.keys()]
        del vectors
        # Create and fit tokenizer to corpus
        tokenizer = EmbeddingAwareTokenizer(
            embedding_vocab=vector_vocab,
            min_counts_common_token=5,
            unknown_token="",  # If no unknown token is provided, will use NLTK-predicted POS tag
            min_subtoken_size=4,
            max_subtokens=3,
        )
        corpus = "\n".join(section_df["section_text"])
        tokenizer.fit(corpus)
        with open(tokenizer_path, "wb") as p:
            pkl.dump(tokenizer, p, pkl.HIGHEST_PROTOCOL)

    print(f"Tokenizing {len(section_df)} sections...")
    section_df["word_tokens"] = section_df["section_text"].map(tokenizer.tokenize)
    section_df.to_pickle(sec_df_path)

    # Calculate section QC/lang metrics
    print("Calculating section-level QC metrics")
    get_char_counts(section_df, pref="section", text_col="section_text")
    count_stopwords(
        section_df,
        pref="section",
        token_col="word_tokens",
        tokenizer=tokenizer.tokenize,
    )
    calc_token_sizes(section_df, pref="section", word_token_col="word_tokens")

    # Save data
    section_df.to_pickle(sec_df_path)

section_df.info()

### Convert to documents
No need to tokenize here since we are just merging tokens in the correct order.

doc_df_path = DATA / "document_corpus.pkl"

if doc_df_path.is_file():
    # del section_df
    doc_df = pd.read_pickle(doc_df_path)
else:
    print("Getting document-level dataframe from section-level dataframe")
    doc_df = sections_df_to_docs(section_df)  # See function/module for details
    del section_df
    # Calculate document QC/lang metrics
    print("Calculating document-level QC metrics")
    get_char_counts(doc_df, pref="doc_main", text_col="clean_text")
    get_char_counts(doc_df, pref="doc_credit", text_col="credit_text")
    count_stopwords(
        doc_df,
        pref="doc_main",
        token_col="clean_word_tokens",
        tokenizer=tokenizer.tokenize,
    )
    count_stopwords(
        doc_df,
        pref="doc_credit",
        token_col="credit_word_tokens",
        tokenizer=tokenizer.tokenize,
    )
    del tokenizer
    calc_token_sizes(doc_df, pref="doc_main", word_token_col="clean_word_tokens")
    calc_token_sizes(doc_df, pref="doc_credit", word_token_col="credit_word_tokens")

    doc_df.to_pickle(DATA / "document_corpus.pkl")

doc_df.info()
